pyproject.toml
[project]
name = "a2ui-line-sample-python"
version = "0.1.0"
description = "A2UI-style agent UI mapped to LINE Flex Messages (Python backend sample)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "pydantic>=2.6",
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
